Configure how to run Picard, not where to find the .jar
Installing Picard on Debian or with Conda installs a "picard" or "picard-tools"
wrapper script which should be used to call Picard instead of invoking the
.jar directly.

This introduces the ability to configure the command with which to run
Picard by setting picard_command in paths.txt. The old way of using
picard_path is still available as a fallback.