Always keep log files
Log files make it possible to find out what has gone wrong when there is a
problem, and they help reproducibility, for example, because they contain
the versions of programs that were run.

Since log files are small, space savings cannot be an argument.

If clutter is a concern, log files could be kept in a separate log/ directory,
but that would be a different change.