Configure logging
Instead of printing stuff with print(), it is a good idea to use a proper
logging infrastructure, such as Python’s logging module that is in the
stdlib. With this, messages can be of different levels (debug, info, warning,
error) and it’s possible to dynamically configure which ones are shown and
where they are written. For example, all logging output could be sent to a
file, but warning-level messages and higher should also be printed to the
screen.