Do not use globals to pass parameters into a function
Using global variables is usually not good style as it makes it hard to reason
about the code. When passing in any parameters that are needed for a
computation explicitly, it is immediately visible where a value comes from.
When globals are used, it is necessary to not only mentally keep track of
the variables in a function, but also of where outside the function they
might have been set.

Example:

def add(x, y):
  global z
  return x + y + z

vs

def add(x, y):
  return x + y

With the first variant, I cannot say what the result will be unless I also
know where z is set and which value it currently has. The second variant can
be understood completely without context.