Move imports to top of file
All imports used in a module should be at the top of the file. It gives a
better overview of the requirements. Import modules within functions that
are called more than once is also ineffecient because the import is re-done
on every call to the function.

Having the imports at the top level will also allow the code to fail quickly
in case an import is not available.

Sometimes, it is useful to have imports within functions, but that is
usually for performance reasons or because the import is only for an
optional feature.