Move description to module docstring
The first string of a module is its docstring (just like with functions). It
is available in the __doc__ variable. I find it nice to have a docstring for
a module (where one would expect it), and to then just re-use it as the
command description in argparse (by passing in description=__doc__).